requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
watchdog>=3.0.0
pillow>=10.0.0
imagehash>=4.3.0
pytesseract>=0.3.10

